apps: separated apps
